Updated column type

+++ database/migrations/2014_10_12_000000_create_users_table.php
@@ -19,7 +19,7 @@ class CreateUsersTable extends Migration
             $table->string('username')->nullable()->unique();
             $table->string('email')->unique();
             $table->string('password');
-            $table->integer('cpf')->nullable();
+            $table->bigInteger('cpf')->nullable();
             $table->boolean('confirmed')->default(false);
             $table->string('email_token')->nullable();
             $table->rememberToken();
